Reorganized c_vector.c and c_vector_ops.c for more visibility

+++ c_vector_capacity.c
@@ -6,14 +6,14 @@
 /*   By: aleger <aleger@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/02/13 16:15:00 by aleger            #+#    #+#             */
-/*   Updated: 2014/02/13 20:01:49 by aleger           ###   ########.fr       */
+/*   Updated: 2014/02/13 22:35:25 by aleger           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "c_vector.h"
 #include <stdlib.h>
 
-int				vector_size(t_vector *self)
+size_t			vector_size(t_vector *self)
 {
 	return (self->count);
 }
